backend/requirements.txt
# FastAPI and server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Pydantic for data validation
pydantic>=2.0.0

# GitHub API client
PyGithub>=2.0.0

# HTTP requests (for fetching PR diffs)
requests>=2.31.0

# Environment variables
python-dotenv>=1.0.0

# Language detection
pygments>=2.0.0

# LLM for CI config inference
openai>=1.0.0
